+++ neo/d3xp/Fx.cpp
@@ -748,8 +748,8 @@ idEntityFx::WriteToSnapshot
 void idEntityFx::WriteToSnapshot( idBitMsg &msg ) const {
 	GetPhysics()->WriteToSnapshot( msg );
 	WriteBindToSnapshot( msg );
-	msg.WriteLong( ( fxEffect != NULL ) ? gameLocal.ServerRemapDecl( -1, DECL_FX, fxEffect->Index() ) : -1 );
-	msg.WriteLong( started );
+	msg.WriteInt( ( fxEffect != NULL ) ? gameLocal.ServerRemapDecl( -1, DECL_FX, fxEffect->Index() ) : -1 );
+	msg.WriteInt( started );
 }
 
 /*
@@ -762,8 +762,8 @@ void idEntityFx::ReadFromSnapshot( const idBitMsg &msg ) {
 
 	GetPhysics()->ReadFromSnapshot( msg );
 	ReadBindFromSnapshot( msg );
-	fx_index = gameLocal.ClientRemapDecl( DECL_FX, msg.ReadLong() );
-	start_time = msg.ReadLong();
+	fx_index = gameLocal.ClientRemapDecl( DECL_FX, msg.ReadInt() );
+	start_time = msg.ReadInt();
 	//#modified-fva; BEGIN
 	start_time = gameLocal.time - (gameLocal.serverTime - start_time);
 	//#modified-fva; END
