+++ neo/d3xp/MultiplayerGame.cpp
@@ -1092,7 +1092,7 @@ void idMultiplayerGame::NewState( gameState_t news, idPlayer *player ) {
 			idBitMsg	matchStartedTimeMsg;
 			byte		matchStartedTimeMsgBuf[ sizeof( matchStartedTime ) ];
 			matchStartedTimeMsg.InitWrite( matchStartedTimeMsgBuf, sizeof( matchStartedTimeMsgBuf ) );
-			matchStartedTimeMsg.WriteLong( matchStartedTime );
+			matchStartedTimeMsg.WriteInt( matchStartedTime );
 			session->GetActingGameStateLobbyBase().SendReliable( GAME_RELIABLE_MESSAGE_MATCH_STARTED_TIME, matchStartedTimeMsg, false );
 
 			fragLimitTimeout = 0;
@@ -1159,7 +1159,7 @@ void idMultiplayerGame::NewState( gameState_t news, idPlayer *player ) {
 			warmupEndTime = gameLocal.serverTime + 1000*cvarSystem->GetCVarInteger( "g_countDown" );
 
 			outMsg.InitWrite( msgBuf, sizeof( msgBuf ) );
-			outMsg.WriteLong( warmupEndTime );
+			outMsg.WriteInt( warmupEndTime );
 			session->GetActingGameStateLobbyBase().SendReliable( GAME_RELIABLE_MESSAGE_WARMUPTIME, outMsg, false );
 
 			// Reset all the scores.
@@ -1936,7 +1936,7 @@ void idMultiplayerGame::WriteToSnapshot( idBitMsg &msg ) const {
 	}
 
 	msg.WriteByte( gameState );
-	msg.WriteLong( nextStateSwitch );
+	msg.WriteInt( nextStateSwitch );
 	msg.WriteShort( currentTourneyPlayer[ 0 ] );
 	msg.WriteShort( currentTourneyPlayer[ 1 ] );
 	for ( i = 0; i < MAX_CLIENTS; i++ ) {
@@ -1973,7 +1973,7 @@ void idMultiplayerGame::ReadFromSnapshot( const idBitMsg &msg ) {
 	}
 
 	newState = (idMultiplayerGame::gameState_t)msg.ReadByte();
-	nextStateSwitch = msg.ReadLong();
+	nextStateSwitch = msg.ReadInt();
 	if ( newState != gameState && newState < STATE_COUNT ) {
 		gameLocal.DPrintf( "%s -> %s\n", GameStateStrings[ gameState ], GameStateStrings[ newState ] );
 		gameState = newState;
@@ -2052,7 +2052,7 @@ void idMultiplayerGame::PlayGlobalSound( int toPlayerNum, snd_evt_t evt, const c
 			if ( !shaderDecl ) {
 				return;
 			}
-			outMsg.WriteLong( gameLocal.ServerRemapDecl( -1, DECL_SOUND, shaderDecl->Index() ) );
+			outMsg.WriteInt( gameLocal.ServerRemapDecl( -1, DECL_SOUND, shaderDecl->Index() ) );
 			type = GAME_RELIABLE_MESSAGE_SOUND_INDEX;
 		} else {
 			outMsg.WriteByte( evt );
@@ -2865,7 +2865,7 @@ void idMultiplayerGame::VoiceChat( const idCmdArgs &args, bool team ) {
 	voiceChatThrottle = gameLocal.realClientTime + 1000;
 
 	outMsg.InitWrite( msgBuf, sizeof( msgBuf ) );
-	outMsg.WriteLong( index );
+	outMsg.WriteInt( index );
 	outMsg.WriteBits( team ? 1 : 0, 1 );
 	session->GetActingGameStateLobbyBase().SendReliableToHost( GAME_RELIABLE_MESSAGE_VCHAT, outMsg );
 }
@@ -2928,7 +2928,7 @@ void idMultiplayerGame::ServerWriteInitialReliableMessages( int clientNum, lobby
 	outMsg.BeginWriting();
 	// send the game state and start time
 	outMsg.WriteByte( gameState );
-	outMsg.WriteLong( matchStartedTime );
+	outMsg.WriteInt( matchStartedTime );
 	outMsg.WriteShort( startFragLimit );
 	// send the powerup states and the spectate states
 	for( int i = 0; i < gameLocal.numClients; i++ ) {
@@ -2948,7 +2948,7 @@ void idMultiplayerGame::ServerWriteInitialReliableMessages( int clientNum, lobby
 		//#modified-fva; BEGIN
 		//outMsg.WriteByte( GAME_RELIABLE_MESSAGE_WARMUPTIME );
 		//#modified-fva; END
-		outMsg.WriteLong( warmupEndTime );
+		outMsg.WriteInt( warmupEndTime );
 		session->GetActingGameStateLobbyBase().SendReliableToLobbyUser( lobbyUserID, GAME_RELIABLE_MESSAGE_WARMUPTIME, outMsg );
 	}
 }
@@ -2961,7 +2961,7 @@ idMultiplayerGame::ClientReadStartState
 void idMultiplayerGame::ClientReadStartState( const idBitMsg &msg ) {
 	// read the state in preparation for reading snapshot updates
 	gameState = (idMultiplayerGame::gameState_t)msg.ReadByte();
-	matchStartedTime = msg.ReadLong( );
+	matchStartedTime = msg.ReadInt( );
 	startFragLimit = msg.ReadShort( );
 
 	int client;
@@ -2990,7 +2990,7 @@ idMultiplayerGame::ClientReadWarmupTime
 ================
 */
 void idMultiplayerGame::ClientReadWarmupTime( const idBitMsg &msg ) {
-	warmupEndTime = msg.ReadLong();
+	warmupEndTime = msg.ReadInt();
 }
 
 /*
@@ -2999,7 +2999,7 @@ idMultiplayerGame::ClientReadWarmupTime
 ================
 */
 void idMultiplayerGame::ClientReadMatchStartedTime( const idBitMsg & msg ) {
-	matchStartedTime = msg.ReadLong();
+	matchStartedTime = msg.ReadInt();
 }
