+++ neo/d3xp/physics/Physics_AF.cpp
@@ -7942,7 +7942,7 @@ void idPhysics_AF::WriteToSnapshot( idBitMsg &msg ) const {
 	int i;
 	idCQuat quat;
 
-	msg.WriteLong( current.atRest );
+	msg.WriteInt( current.atRest );
 	msg.WriteFloat( current.noMoveTime );
 	msg.WriteFloat( current.activateTime );
 	msg.WriteDeltaFloat( 0.0f, current.pushVelocity[0], AF_VELOCITY_EXPONENT_BITS, AF_VELOCITY_MANTISSA_BITS );
@@ -7989,7 +7989,7 @@ void idPhysics_AF::ReadFromSnapshot( const idBitMsg &msg ) {
 	int i, num;
 	idCQuat quat;
 
-	current.atRest = msg.ReadLong();
+	current.atRest = msg.ReadInt();
 	current.noMoveTime = msg.ReadFloat();
 	current.activateTime = msg.ReadFloat();
 	current.pushVelocity[0] = msg.ReadDeltaFloat( 0.0f, AF_VELOCITY_EXPONENT_BITS, AF_VELOCITY_MANTISSA_BITS );
