+++ neo/idlib/Parser.h
@@ -226,9 +226,9 @@ private:
 	int				Directive_ifndef();
 	int				Directive_else();
 	int				Directive_endif();
-	int				EvaluateTokens( idToken *tokens, signed long int *intvalue, double *floatvalue, int integer );
-	int				Evaluate( signed long int *intvalue, double *floatvalue, int integer );
-	int				DollarEvaluate( signed long int *intvalue, double *floatvalue, int integer);
+	int				EvaluateTokens( idToken *tokens, int *intvalue, double *floatvalue, int integer );
+	int				Evaluate( int *intvalue, double *floatvalue, int integer );
+	int				DollarEvaluate( int *intvalue, double *floatvalue, int integer);
 	int				Directive_define();
 	int				Directive_elif();
 	int				Directive_if();
